Fix linking readline from non-default library directories
pkg_search_module(READLINE readline) returns only "-lreadline" in
READLINE_LIBRARIES. If readline is located in a library directory that
the is not searched by default by the linker, this by itself is not
enough information to successfully link foma-bin.

pkg_search_module also provides a …_LIBRARY_DIRS variable, which
contains the directories that need to be in the library search path in
order for the library to be found. Pass it to target_link_directories
for foma-bin.